Add renderables for f3df, update types in i3df
and add statistics.

+++ i3df/src/lib.rs
@@ -52,8 +52,8 @@ pub struct SectorHeader {
     pub magic_bytes: u32,
     pub format_version: u32,
     pub optimizer_version: u32,
-    pub sector_id: u64,
-    pub parent_sector_id: Option<u64>,
+    pub sector_id: usize,
+    pub parent_sector_id: Option<usize>,
 
     pub bbox_min: [f32; 3],
     pub bbox_max: [f32; 3],
@@ -199,10 +199,10 @@ pub fn parse_sector_header(mut input: &mut impl BufRead) -> Result<SectorHeader,
     assert_eq!(format_version, 7);
     let optimizer_version = input.read_u32::<LittleEndian>()?;
 
-    let sector_id = input.read_u64::<LittleEndian>()?;
+    let sector_id = input.read_u64::<LittleEndian>()? as usize;
     let parent_sector_id = match input.read_u64::<LittleEndian>()? {
         std::u64::MAX => None,
-        x => Some(x),
+        x => Some(x as usize),
     };
 
     let bbox_min_x = input.read_f32::<LittleEndian>()?;
